Sprint 4+5

+++ passenger-service/src/main/java/com/example/core/service/PassengerService.java
@@ -5,6 +5,7 @@ import com.example.core.dto.PassengerMapper;
 import com.example.core.entity.Passenger;
 import com.example.core.exception.ResourceNotFoundException;
 import com.example.core.repository.PassengerRepository;
+import io.micrometer.tracing.annotation.NewSpan;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
@@ -18,12 +19,12 @@ import java.util.Optional;
 @Service
 public class PassengerService {
     private static final Logger log = LoggerFactory.getLogger(PassengerService.class);
-    @Autowired
-    private PassengerRepository passengerRepository;
-    @Autowired
-    private PassengerMapper passengerMapper;
+
+    private final PassengerRepository passengerRepository;
+    private final PassengerMapper passengerMapper;
 
     @Transactional
+    @NewSpan("register-passenger")
     public PassengerDto registerPassenger(PassengerDto passengerDto) {
         log.info("Registering passenger: {}", passengerDto);
         Passenger passenger = passengerMapper.toEntity(passengerDto);
@@ -33,10 +34,10 @@ public class PassengerService {
         return passengerMapper.toDto(savedPassenger);
     }
 
-    public Optional<PassengerDto> findPassengerByEmail(String email) {
+   /* public Optional<PassengerDto> findPassengerByEmail(String email) {
         Optional<Passenger> passenger = passengerRepository.findByEmail(email);
         return passenger.map(passengerMapper::toDto);
-    }
+    }*/
 
     @Transactional
     public PassengerDto updatePassenger(PassengerDto passengerDto) {
@@ -49,6 +50,10 @@ public class PassengerService {
 
     @Transactional
     public void deletePassenger(Long id) {
+        if (!passengerRepository.existsById(id)) {
+            log.error("Attempted to delete non-existing passenger with ID {}", id);
+            throw new ResourceNotFoundException("Passenger not found");
+        }
         passengerRepository.deleteById(id);
     }
